requirements.txt
Flask
gunicorn
google-generativeai
python-dotenv
Flask-SQLAlchemy
psycopg2-binary